fix: fix amounts being strings in neo4j db
The datatype float is not native to the neo4j database, so neo4j turns
floats into strings. Fix decimals by changing datatype from float to
double in code.
Remove unecessary null parameter for requirement since the neo4j
overwrites the id when it saves the requirement.

+++ src/main/java/practice/recipebase/interpreter/Measurement.java
@@ -1,4 +1,4 @@
 package practice.recipebase.interpreter;
 
-public record Measurement(String unit, Float amount) {
+public record Measurement(String unit, Double amount) {
 }
